refactor:convert @electron-forge/template-typescript-webpack to use BaseTemplate
