+++ packages/template/typescript-webpack/package.json
@@ -16,6 +16,10 @@
   "dependencies": {
     "@electron-forge/async-ora": "6.0.0-beta.46",
     "@electron-forge/shared-types": "6.0.0-beta.46",
+    "@electron-forge/template-base": "6.0.0-beta.46",
     "fs-extra": "^8.1.0"
+  },
+  "devDependencies": {
+    "@electron-forge/test-utils": "6.0.0-beta.46"
   }
 }
